START OF THE END :-/ DJANOG MAKE FUN OF US WITH THIS BUG ... WE WILL FIX IT

+++ order/models.py
@@ -59,8 +59,8 @@ class Order(BaseModel):
 class OrderItem(BaseModel):
     order = models.ForeignKey(Order, on_delete=models.PROTECT, related_name='items')
     product = models.ForeignKey(Food, on_delete=models.PROTECT)
-    price = models.DecimalField(max_digits=10, decimal_places=3)
-    quantity = models.PositiveIntegerField(default=1)
+    price = models.DecimalField(max_digits=10, decimal_places=4)
+    quantity = models.DecimalField(max_digits=5, decimal_places=2, default=1)
 
     def __str__(self):
         return f"{self.quantity} x {self.product.name} in Order #{self.order.id}"
